perf: stop shipping full Pokémon dataset to client; fix React/lint issues
Split the client bundle from the ~2.9MB game dataset and clean up
correctness/lint issues surfaced by eslint and a production build.

Performance
- Move server-only `resolveShareCode` into `src/lib/share-resolve.ts` so
  the client-safe `share.ts` (used by ShareButton) no longer transitively
  imports the `@/data` barrel. This removes a 2.9MB data chunk from the
  /history and /play client bundles. Total client JS: 3914KB -> 928KB.
- Memoize the history migration pass with useMemo.
- Make icon-192/icon-512 static (drop edge runtime, force-static) so they
  prerender instead of rendering per request.

Correctness / cleanup
- Fix "ref accessed during render" in game-client (use state.team in render;
  sync teamRef in an effect for deferred setTimeout closures).
- Type history migration helpers with LegacyTeamHistoryEntry/Pokemon + a
  type guard instead of `any`.
- Remove dead code (getOverlappingTeamIndices, isLineOnTeam, StatChart
  primaryType prop) and unused imports.
- Replace the create-next-app boilerplate README with real project docs.

eslint clean (was 7 errors + 4 warnings); `next build` passes.
