+++ src/app/icon-512/route.tsx
@@ -1,6 +1,6 @@
 import { ImageResponse } from "next/og";
 
-export const runtime = "edge";
+export const dynamic = "force-static";
 
 export async function GET() {
   return new ImageResponse(
